REF: dtypes.astype
