+++ pandas/core/internals/blocks.py
@@ -35,8 +35,8 @@ from pandas.util._decorators import cache_readonly
 from pandas.util._exceptions import find_stack_level
 from pandas.util._validators import validate_bool_kwarg
 
+from pandas.core.dtypes.astype import astype_array_safe
 from pandas.core.dtypes.cast import (
-    astype_array_safe,
     can_hold_element,
     find_common_type,
     infer_dtype_from,
